refactor: extract the MPD check in a module
Continue the work to created closely related modules instead of a big
one.
